fix: missing FMC questions
