+++ packages/web-app/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@expanse/web-app",
-  "version": "1.0.18",
+  "version": "1.0.20",
   "private": true,
   "dependencies": {
     "@progress/kendo-licensing": "1.3.5",
